refactor: move screen classes to view package
